+++ src/web/wizard/stores/AlertRuleStore.js
@@ -114,9 +114,10 @@ const AlertRuleStore = Reflux.createStore({
         const promise = fetch('POST', url, request)
             .then(() => {
                 UserNotification.success('Stream successfully created, Alert condition successfully created, Alert notification successfully created');
+                // TODO is this necessary? Try to remove
                 this.list();
                 return true;
-            }, (error) => {
+            }, error => {
                 UserNotification.error(`Creating alert rule failed with status: ${error.message}`,
                     'Could not create alert rule');
             });